chore: return both session and engine and test both

+++ backend/src/db/models/fighters.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, Float
 from .base import Base
 
 class Fighter(Base):
@@ -6,4 +6,4 @@ class Fighter(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     name = Column(String, unique=True, nullable=False)
     plays = Column(Integer, nullable=False)
-    winrate = Column(Integer, nullable=False)
+    winrate = Column(Float, nullable=False)
